pyproject.toml
[project]
name = "telegramjobnotifier"
version = "0.1.0"
description = "A Telegram bot that forwards messages from admin to channels"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.27.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
